move src/cdp/websocketTransport.ts -> src/utils/websocketTransport.ts
Closes #544
